+++ src/main/java/step5/utils/StringUtils.java
@@ -1,4 +1,4 @@
-package step3.utils;
+package step5.utils;
 
 import java.util.*;
