+++ src/main/java/step5/winning/WinningPrize.java
@@ -1,4 +1,4 @@
-package step3.winning;
+package step5.winning;
 
 import java.util.Arrays;
