+++ src/main/java/step5/winning/WinningStatistics.java
@@ -1,9 +1,9 @@
-package step3.winning;
+package step5.winning;
 
-import step3.lotto.Lotto;
-import step3.lotto.LottoCount;
-import step3.lotto.LottoNumber;
-import step3.lotto.LottoTicket;
+import step5.lotto.Lotto;
+import step5.lotto.LottoCount;
+import step5.lotto.LottoNumber;
+import step5.lotto.LottoTicket;
 
 import java.math.*;
 import java.util.*;
