+++ src/test/java/step5/utils/StringUtilsTest.java
@@ -1,7 +1,8 @@
-package step3.utils;
+package step5.utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
