+++ src/test/java/step5/winning/WinningNumbersTest.java
@@ -1,9 +1,10 @@
-package step3.winning;
+package step5.winning;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 class WinningNumbersTest {
